Renamed dashboard file and formatting changes
