+++ sinks/dashboard/dashboard.py
@@ -1,56 +1,94 @@
-import os
-import sys
-import json
+import os, sys, json
 
+from pyqtgraph.Qt.QtCore import Qt, QTimer
+from pyqtgraph.Qt.QtGui import QPainter
+from pyqtgraph.Qt.QtWidgets import (
+    QGraphicsView,
+    QGraphicsScene,
+    QApplication,
+    QWidget,
+    QMenuBar,
+    QVBoxLayout,
+    QGraphicsItem,
+    QGraphicsRectItem
+)
 from items import registry
-from pyqtgraph.Qt import QtCore
-import pyqtgraph as pg
-from pyqtgraph.Qt import QtWidgets
-from pyqtgraph.dockarea.Dock import Dock
-from pyqtgraph.dockarea.DockArea import DockArea
-from items.plot_dash_item import PlotDashItem
-from items.can_message_table import CanMsgTableDashItem
 from omnibus.util import TickCounter
 from utils import prompt_user
 
+# These need to be imported to be added to the registry
+from items.plot_dash_item import PlotDashItem
+from items.can_message_table import CanMsgTableDashItem
 
-class Dashboard(QtWidgets.QWidget):
-    """
-    Displays a grid of plots in a window
-    """
+# Custom class derived from QGraphicsView to capture mouse
+# wheel events by overriding the wheelEvent function
+class MyQGraphicsView(QGraphicsView):
+    def __init__ (self, parent=None):
+        # Initialize the super class
+        super(MyQGraphicsView, self).__init__ (parent)
 
+    def wheelEvent(self, event):
+        # Zoom if Shift is held, otherwise scroll
+        if event.modifiers() & Qt.ShiftModifier:
+            # Zoom Factor
+            zoomInFactor = 1.1
+            zoomOutFactor = 1 / zoomInFactor
+
+            # Zooms to the position of the mouse
+            self.setTransformationAnchor(QGraphicsView.AnchorUnderMouse)
+
+            # Determine zoom in/out by how much the wheel moves
+            angle = event.angleDelta()
+            if angle.x() > 0 or angle.y() > 0:
+                zoomFactor = zoomInFactor
+            elif angle.x() < 0 or angle.y() > 0:
+                zoomFactor = zoomOutFactor
+            else:
+                zoomFactor = 1
+
+            self.scale(zoomFactor, zoomFactor)
+        else:
+            super(MyQGraphicsView, self).wheelEvent(event)
+
+
+# Custom Dashboard class derived from QWidget
+class Dashboard(QWidget):
     def __init__(self, callback):
-        # Initilize Super Class
-        QtWidgets.QWidget.__init__(self)
+        # Initialize the super class
+        super().__init__()
 
-        # called every frame to get new data
-        self.callback = callback
+        # Called every frame to get new data
+        self.callback = callback 
+
+        # Dictionary to map rectitems to widgets 
+        # and dashitems
+        self.widgets = {}
+
+        # Keep track of if editing is allowed
+        self.locked = False
 
         # The file from which the dashboard is loaded
         self.filename = "savefile.json"
         self.filename_cache = [self.filename]
 
-        # A list of all docks in the dock area
-        # doubles as a list of all the items in
-        # the dashboare by making use of
-        # item = dock.widgets[0]
-        self.docks = []
-
-        # Create the QT GUI which will be the dashboard
-        # To do, find a way to move all of this to
-        # another file
-        ###############################################
+        # Create a GUI
+        self.width = 1100
+        self.height = 700
         self.setWindowTitle("Omnibus Dashboard")
-        self.resize(1000, 600)
+        self.resize(self.width, self.height)
 
-        # Create GridLayout, will be
-        # Adding components to this as
-        # time goes on
-        self.layout = QtWidgets.QVBoxLayout()
-        self.setLayout(self.layout)
+        # Create a large scene underneath the view
+        self.scene = QGraphicsScene(0, 0, self.width*100, self.height*100)
 
-        # Add a menu bar to the layout
-        menubar = QtWidgets.QMenuBar(self)
+        # Create a grid layout
+        self.layout = QVBoxLayout()
+
+        # Create a menubar for actions
+        menubar = QMenuBar(self)
+
+        # List to keep track of menu bar action that 
+        # can be disabled when dashboard is locked
+        self.disable = []
 
         # Create a sub menu which will be used
         # to add items to our dash board.
@@ -73,147 +111,163 @@ class Dashboard(QtWidgets.QWidget):
         for i in range(len(registry.get_items())):
             new_action = add_item_menu.addAction(registry.get_items()[i].get_name())
             new_action.triggered.connect(prompt_and_add(i))
+            self.disable.append(new_action)
 
         # Add an action to the menu bar to save the
         # layout of the dashboard.
         add_save_menu = menubar.addMenu("Save")
         save_layout_action = add_save_menu.addAction("Save Current Config")
         save_layout_action.triggered.connect(self.save)
+        self.disable.append(save_layout_action)
 
         # Add an action to the menu bar to load the
         # layout of the dashboard.
         add_restore_menu = menubar.addMenu("Load")
         restore_layout_action = add_restore_menu.addAction("Load from File")
         restore_layout_action.triggered.connect(self.load)
+        self.disable.append(restore_layout_action)
 
         # Add an action to the menu bar to open a file
         add_open_menu = menubar.addMenu("Open")
         open_file_action = add_open_menu.addAction("Open File")
         open_file_action.triggered.connect(self.switch)
+        self.disable.append(open_file_action)
+
+        # Add an action to the menu bar to lock/unlock
+        # the dashboard
+        add_open_menu = menubar.addMenu("Lock")
+        lock_action = add_open_menu.addAction("Lock Dashboard")
+        lock_action.triggered.connect(self.lock)
+        self.disable.append(lock_action)
+        unlock_action = add_open_menu.addAction("Unlock Dashboard")
+        unlock_action.triggered.connect(self.unlock)
 
         self.layout.setMenuBar(menubar)
 
-        # Create the Dock Area which will house all of the items
-        self.area = DockArea()
-        self.layout.addWidget(self.area)
-
-        # Restore last save state
-        self.load()
-
+        # Set the counter
         self.counter = TickCounter(1)
 
+        # Create the view and add it to the widget
+        self.view = MyQGraphicsView(self.scene)
+        self.view.setDragMode(QGraphicsView.ScrollHandDrag)
+        self.view.setRenderHints(QPainter.Antialiasing)
+        self.layout.addWidget(self.view)
+        self.setLayout(self.layout)
+
+    # Method to add widgets
+    def add(self, dashitem, pos=None):
+        # Add the dash item to the scene and get
+        # its proxy widget and dimension
+        proxy = self.scene.addWidget(dashitem)
+        height = proxy.size().height()
+        width = proxy.size().width()
+
+        # If a position is given, use it. If not, position
+        # the widget in the center of the current view
+        if pos:
+            mapped = self.view.mapToScene(pos[0], pos[1])
+            xpos = mapped.x()
+            ypos = mapped.y()
+        else:
+            # Get the current size of the view area and map
+            # it to the underlying scene
+            viewport = self.view.viewport().size()
+            view_xpos = viewport.width()/2
+            view_ypos = viewport.height()/2
+
+            mapped = self.view.mapToScene(view_xpos, view_ypos)
+
+            # Center the widget in the view. Qt sets position
+            # based on the upper left corner, so subtract
+            # half the width and height of the widget to 
+            # center the center
+            xpos = mapped.x() - (width/2)
+            ypos = mapped.y() - (height/2)
+
+        proxy.setPos(xpos, ypos)
+        proxy.setFocusPolicy(Qt.NoFocus)
+        
+        # Create a rectangle around the proxy widget
+        # to make it movable and selectable
+        rect = QGraphicsRectItem(xpos-1, ypos-1, width+1, height+1)
+        proxy.setParentItem(rect)
+        rect.setFlag(QGraphicsItem.GraphicsItemFlag.ItemIsMovable)
+        rect.setFlag(QGraphicsItem.GraphicsItemFlag.ItemIsSelectable)
+        self.scene.addItem(rect)
+
+        # Map the proxy widget and dashitem to the rectitem
+        self.widgets[rect] = [proxy, dashitem]
+
+    # Method to remove a widget
+    def remove(self, item):
+        # Remove the rectangle from the scene,
+        # delete the proxy widget and tell the
+        # dashitem to stop sending data
+        components = self.widgets[item]
+        proxy = components[0]
+        dashitem = components[1]
+        self.scene.removeItem(item)
+        proxy.deleteLater()
+        dashitem.on_delete()
+
+    # Method to remove all widgets
+    def remove_all(self):
+        for item in self.widgets:
+            self.remove(item)
+
+        # Clear the mapping list. This can't be
+        # done while iterating through the dict
+        # because it changes the length and causes
+        # a RunTime Error
+        self.widgets = {}
+
+    # Method to load layout from file
     def load(self):
-        # First, we need to clear all the
-        # docks currently on the screen.
+        # First remove all the current widgets
+        self.remove_all()
 
-        # First, make sure that the dash
-        # items within the docks are not
-        # Registered with any series
-        for dock in self.docks:
-            item = dock.widgets[0]
-            item.on_delete()
-
-        # Second, remove the entire dock area,
-        # thereby deleting all of the docks
-        # contained within
-        self.layout.removeWidget(self.area)
-        self.area.deleteLater()
-        del self.area
-
-        # Now we recreate the dock area,
-        # ahearing to the layout specified in
-        # the save file
-
-        # Create a new dock area
-        self.area = DockArea()
-        self.layout.addWidget(self.area)
-
-        # re-populate our dash area using
-        # the save file
-        self.docks = []
-
-        # if the save file exists, set our
-        # data variable to it
-        if os.path.isfile(self.filename):
-            with open(self.filename, 'r') as savefile:
+        # Then load the data from the savefile
+        if os.path.exists(self.filename):
+            with open(self.filename, "r") as savefile:
                 data = json.load(savefile)
         else:
-            data = {
-                "items": [],
-                "layout": None
-            }
+            return
 
-        # for every item specified by the save file,
-        # add that item back to the dock
-        for i, item in enumerate(data["items"]):  # { 0: {...}, 1: ..., ...}
-            # Convert the name of the class back into the class
-            # See save() method
-            for item_type in registry.get_items():
-                if item["class"] == item_type.get_name():
-                    self.add(item_type(item["props"]))
-                    break
-
-        # restore the layout
-        if data["layout"]:
-            self.area.restoreState(data["layout"])
-
-    def save(self):
-        data = {
-            "items": [],
-            "layout": None
-        }
-        # The way that self.area.saveState() works
-        # is by assigning properties to each dock
-        # base on its name. This means we need to
-        # set the names in such a way that when
-        # the docks are re-added, the names align
-        # therefore, we set the name of each dock
-        # to its index within docks
-        for i in range(len(self.docks)):
-            self.docks[i]._name = str(i)
-            self.docks[i].label.setText(str(i))
-
-        # store layout data to data["layout"]
-        data["layout"] = self.area.saveState()
-
-        # store data on the specific dash items to
-        # data["items"]
-        data["items"] = []
-        for dock in self.docks:
-            item = dock.widgets[0]
-
+        # Add every widget in the data
+        for widget in data:
             # ObjectTypes can't be converted to JSON
-            # Take the name of the class instead
+            # See the save method
             for item_type in registry.get_items():
-                if type(item) == item_type:
-                    data["items"].append({"props": item.get_props(), "class": item_type.get_name()})
+                if widget["class"] == item_type.get_name():
+                    self.add(item_type(widget["props"]), widget["pos"])
                     break
 
-        # Save to the save file
-        with open(self.filename, 'w') as savefile:
+    # Method to save current layout to file
+    def save(self):
+        data = []
+        for items in self.widgets.values():
+            # Get the proxy widget and dashitem
+            proxy = items[0]
+            dashitem = items[1]
+
+            # Get the coordinates of the proxy widget
+            # on the view not the scene
+            scenepos = proxy.scenePos()
+            viewpos = self.view.mapFromScene(scenepos)
+
+            # Add the position, dashitem name and dashitem props
+            for item_type in registry.get_items():
+                if type(dashitem) == item_type:
+                    data.append({"class": item_type.get_name(), 
+                                 "props": dashitem.get_props(),
+                                 "pos": [viewpos.x(), viewpos.y()]})
+                    break
+
+        # Write data to savefile
+        with open(self.filename, "w") as savefile:
             json.dump(data, savefile)
 
-    def add(self, dashitem):
-        # Create a new dock to be added to the dock area
-        dock = Dock(f"{len(self.docks)}", closable=True)
-
-        # Create a call back to execute when docks close to ensure cleaning up is done
-        # correctly
-        def custom_callback(dock_arg):
-            dashitem.on_delete()
-            self.docks = [dock for dock in self.docks if dock.widgets[0] != dashitem]
-
-        dock.sigClosed.connect(custom_callback)
-
-        # add widget to dock
-        dock.addWidget(dashitem)
-
-        # add dock to dock area
-        self.area.addDock(dock, 'right')
-
-        # add dock to dock list
-        self.docks.append(dock)
-
+    # Method to switch to a layout in a different file
     def switch(self):
         self.save()
         filename = prompt_user(
@@ -231,31 +285,62 @@ class Dashboard(QtWidgets.QWidget):
         # If the filename entered is not valid
         # this exhibits the behaviour of creating
         # a new one
-
         if filename not in self.filename_cache:
             self.filename_cache.append(filename)
 
         self.filename = filename
         self.load()
 
-    # called every frame
+    # Method to lock dashboard
+    def lock(self):
+        self.locked = True
+        self.setWindowTitle("Omnibus Dashboard - LOCKED")
+
+        # Disable menu actions
+        for menu_item in self.disable:
+            menu_item.setEnabled(False)
+
+        # Disable selecting and moving plots
+        for rect in self.widgets:
+            rect.setFlag(QGraphicsItem.GraphicsItemFlag.ItemIsMovable, enabled=False)
+            rect.setFlag(QGraphicsItem.GraphicsItemFlag.ItemIsSelectable, enabled=False)
+        
+    # Method to unlock dashboard
+    def unlock(self):
+        self.locked = False
+        self.setWindowTitle("Omnibus Dashboard")
+
+        # Enable menu actions
+        for menu_item in self.disable:
+            menu_item.setEnabled(True)
+
+        # Enable selecting and moving plots
+        for rect in self.widgets:
+            rect.setFlag(QGraphicsItem.GraphicsItemFlag.ItemIsMovable, enabled=True)
+            rect.setFlag(QGraphicsItem.GraphicsItemFlag.ItemIsSelectable, enabled=True)
+        
+    # Method to get new data for widgets
     def update(self):
         self.counter.tick()
-
-        # Filter to 5 frames per update on analytics
-        if not (self.counter.tick_count() % 5):
-            fps = self.counter.tick_rate()
-
         self.callback()
 
+    # Method to capture key presses
+    def keyPressEvent(self, event):
+        if event.key() == Qt.Key_Backspace and not self.locked:
+            for item in self.scene.selectedItems():
+                self.remove(item)
+                self.widgets.pop(item)
 
+
+# Function to launch the dashboard
 def dashboard_driver(callback):
-    app = QtWidgets.QApplication(sys.argv)
+    app = QApplication(sys.argv)
     dash = Dashboard(callback)
 
-    timer = QtCore.QTimer()
+    timer = QTimer()
     timer.timeout.connect(dash.update)
     timer.start(16)  # Capped at 60 Fps, 1000 ms / 16 ~= 60
 
     dash.show()
-    sys.exit(app.exec_())
+    dash.load()
+    app.exec()
