+++ sinks/dashboard/main.py
@@ -1,7 +1,7 @@
 from omnibus import Receiver
 
 import parsers
-from dashboard2 import dashboard_driver
+from dashboard import dashboard_driver
 
 receiver = Receiver("")  # subscribe to all channels
